+++ ui/src/pages/namespace/Mirror/Syncs/Detail/ActivityDetail.tsx
