+++ ui/src/pages/namespace/Mirror/Syncs/Detail/Header/index.tsx
