+++ ui/src/pages/namespace/Mirror/Syncs/Detail/Logs/Entry.tsx
