+++ ui/src/pages/namespace/Mirror/Syncs/Detail/Logs/ScrollContainer.tsx
