+++ ui/src/pages/namespace/Mirror/Syncs/Detail/Logs/index.tsx
