+++ ui/src/pages/namespace/Mirror/Syncs/Detail/index.tsx
