+++ ui/src/pages/namespace/Mirror/Syncs/Header/EditDialog.tsx
