+++ ui/src/pages/namespace/Mirror/Syncs/Header/SyncDialog.tsx
