+++ ui/src/pages/namespace/Mirror/Syncs/Row/index.tsx
